move projection code to vcm/calc/flux.py
