+++ external/vcm/vcm/calc/__init__.py
@@ -2,6 +2,11 @@ from .calc import apparent_heating, apparent_source, local_time, weighted_averag
 from .metrics import r2_score
 from .q_terms import compute_Q_terms
 from ._zenith_angle import cos_zenith_angle
+from .flux import (
+    convergence_cell_center,
+    convergence_cell_interface,
+    project_vertical_flux,
+)
 
 __all__ = [
     "storage_and_advection",
